diffuse/specular blend support
however, the monte-carlo integration is b0rked; it's always a 50/50
split between diffuse and specular rays if both are required by the
material, so e.g. a material with 1% diffuse and 99% specular sounds
vastly different from 0% diffuse and 100% specular. I need to be able to
analyze how much energy is reflected by either type and make a better
random pick based on that? hm!